General default gateway client improvements
- Upgrade implementations to a new retryWhen factory
- Backoff/jitter in ReconnectOptions deprecated due to new retry factory
- Improve resume with zero initial backoff in DefaultGatewayClient
- Improve state management, remove some atomic fields
- Change GatewayClient::isConnected return type to Mono<Boolean>
- Some reconnect events have changed their return types to long
- Internally use Mono-based PayloadHandlers
- Add new stateEvents API to GatewayClient
- Add dedicated type for invalid sessions
- Make ReconnectException extends GatewayException

+++ core/src/main/java/discord4j/core/event/domain/lifecycle/ReconnectEvent.java
@@ -27,9 +27,9 @@ import discord4j.gateway.ShardInfo;
  */
 public class ReconnectEvent extends GatewayLifecycleEvent {
 
-    private final int currentAttempt;
+    private final long currentAttempt;
 
-    public ReconnectEvent(GatewayDiscordClient gateway, ShardInfo shardInfo, int currentAttempt) {
+    public ReconnectEvent(GatewayDiscordClient gateway, ShardInfo shardInfo, long currentAttempt) {
         super(gateway, shardInfo);
         this.currentAttempt = currentAttempt;
     }
@@ -39,7 +39,7 @@ public class ReconnectEvent extends GatewayLifecycleEvent {
      *
      * @return The current reconnect attempt.
      */
-    public int getCurrentAttempt() {
+    public long getCurrentAttempt() {
         return currentAttempt;
     }
